Add support to update corpus and trace package code
This commit adds a script to update the corpus with the
list of installed packages. This can be invoked with the
make rule `udpate-corpus`.

A rule `install-dependencies` has been added to install all
the dependencies.

The rule `trace` has been updated to run the tracer on
vignettes, examples and test cases extracted from the the
packages in the corpus.

The tracer is invoked via GNU Parallel. Parallelization is
controlled by a file which specifies the number of parallel
tracer instances that should be run.